Added code to ensure that assets are not copied when extra assets paths in Android are not specified

+++ extras/Projucer/Source/Project Saving/jucer_ProjectExport_Android.h
@@ -218,12 +218,16 @@ public:
 
         writeCmakeFile (appFolder.getChildFile ("CMakeLists.txt"));
 
-        File extraAssets (getProject().getFile().getParentDirectory().getChildFile(androidExtraAssetsFolder.get()));
-        if (extraAssets.exists() && extraAssets.isDirectory())
+        const String androidExtraAssetsFolderValue = androidExtraAssetsFolder.get();
+        if (androidExtraAssetsFolderValue.isNotEmpty())
         {
-            const File assetsFolder (appFolder.getChildFile ("src/main/assets"));
-            if (assetsFolder.deleteRecursively())
-                extraAssets.copyDirectoryTo (assetsFolder);
+            File extraAssets (getProject().getFile().getParentDirectory().getChildFile(androidExtraAssetsFolderValue));
+            if (extraAssets.exists() && extraAssets.isDirectory())
+            {
+                const File assetsFolder (appFolder.getChildFile ("src/main/assets"));
+                if (assetsFolder.deleteRecursively())
+                    extraAssets.copyDirectoryTo (assetsFolder);
+            }
         }
     }
